GW-8: Add Extensions and Dao Tests

+++ app/src/main/java/com/greyp/weather/data/remote/responses/Snow.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Snow(
     @SerializedName(value = "1h")
-    val volumeIn1h: Int,
+    val volumeIn1h: Double,
     @SerializedName(value = "3h")
-    val volumeIn3h: Int
+    val volumeIn3h: Double
 )
